Split PartitionBasedMvRefreshProcessorTest.java file
